Cambio las rutas a web.js
