+++ backend/controllers/ChatController.js
@@ -1,5 +1,84 @@
 const client = require("../Database/RedisClient");
 
+async function create(req, res) {
+  try {
+    const { user1_ID, user2_ID } = req.body;
+
+    // Verificar si los usuarios ya están en una sala de chat juntos
+    const existingChatID = await findChatByUserIDs(user1_ID, user2_ID);
+    if (existingChatID) {
+      return res
+        .status(400)
+        .json({ message: "Los usuarios ya están en una sala de chat juntos" });
+    }
+
+    if (!user1_ID || !user2_ID) {
+      return res
+        .status(400)
+        .json({ message: "user1_ID y user2_ID son obligatorios" });
+    }
+
+    // Crear la sala de chat
+    const chatData = { user1_ID, user2_ID };
+    await addChatToList(chatData); // Llamar a la función para agregar chat a la lista
+
+    res.status(201).json({ message: "Chat creado correctamente" });
+  } catch (error) {
+    console.error("Error al crear chat:", error);
+    res.status(500).json({ message: "Error interno del servidor" });
+  }
+}
+
+async function getChatsFromUser(req, res) {
+  try {
+    const { user_ID } = req.body;
+
+    const userChats = await getUserChats(user_ID);
+
+    res.status(200).json({ userChats });
+  } catch (err) {
+    res.status(500).json({ message: "Error interno del servidor" });
+    console.log("Error al traer los chats del usuario:", err);
+  }
+}
+
+async function openChat(req, res) {
+  try {
+    const { chat_ID } = req.body;
+
+    const messagesFromChat = await getMessagesFromChat(chat_ID);
+
+    res.status(200).json({ messagesFromChat });
+  } catch (err) {
+    res.status(500).json({ message: "Error interno del servidor" });
+    console.log("Error al traer el chat del usuario:", err);
+  }
+}
+
+async function sendMessage(req, res) {
+  try {
+    const { sender, receiver, content, chat_ID } = req.body;
+
+    const messageData = { sender, receiver, content, chat_ID };
+
+    const messageJSON = JSON.stringify(messageData);
+
+    // Agregar el chat al final de la lista
+    await client.rPush("messages_list", messageJSON, (err, reply) => {
+      if (err) {
+        reject(err);
+      } else {
+        resolve(reply);
+      }
+    });
+
+    res.status(201).json({ message: "Mensaje enviado correctamente" });
+  } catch (err) {
+    res.status(500).json({ message: "Error interno del servidor" });
+    console.log("Error al enviar mensaje:", err);
+  }
+}
+
 async function findChatByUserIDs(user1_ID, user2_ID) {
   // Obtener todas las salas de chat
   const allChats = await client.lRange("chats_list", 0, -1);
@@ -51,35 +130,16 @@ async function getUserChats(user_ID) {
   return allUserChats.length !== 0 ? allUserChats : null;
 }
 
-// Función para guardar mensajes
-async function sendMessage(message) {
-  const messageJSON = JSON.stringify(message);
-
-  // Agregar el chat al final de la lista
-  await client.rPush("messages_list", messageJSON, (err, reply) => {
-    if (err) {
-      reject(err);
-    } else {
-      resolve(reply);
-    }
-  });
-}
-
 //Funcion para recoger los mensajes de un chat
 async function getMessagesFromChat(chat_ID) {
-
   const allMessages = await client.lRange("messages_list", 0, -1);
-  console.log(allMessages);
 
   const messagesFromChat = [];
 
-
   for (const messageJSON of allMessages) {
-    console.log(messageJSON);
     const message = JSON.parse(messageJSON);
-    console.log(message);
 
-    console.log(message.chat_ID+ " "+ chat_ID);
+    console.log(message.chat_ID + " " + chat_ID);
     if (message.chat_ID == chat_ID) {
       messagesFromChat.push(message);
     }
@@ -89,9 +149,8 @@ async function getMessagesFromChat(chat_ID) {
 }
 
 module.exports = {
-  findChatByUserIDs,
-  addChatToList,
-  getUserChats,
-  sendMessage,
-  getMessagesFromChat,
+  create,
+  getChatsFromUser,
+  openChat,
+  sendMessage
 };
